Fix varnish assertion logic
We can't delete the request data from the stash just after
making a request. The "page is cached by varnish" assertion
in fact needs to replay the request to check the X-Varnish: header,
so it needs all the original headers/cookies/etc to correctly
replay the request.

We don't force removal of the request info from the "stash",
so the reset_stash() call in Test/Cukes.pm does everything.